refactor(all): 도메인 확실하게 분류 및 패키지 정리
member, reservation, slice, stadium 이렇게 4개의 도메인으로 분류
