+++ src/main/java/com/example/tuktuk/domain/slice/SliceId.java
@@ -1,4 +1,4 @@
-package com.example.tuktuk.domain.event;
+package com.example.tuktuk.domain.slice;
 
 import jakarta.persistence.Column;
