+++ src/main/java/com/example/tuktuk/domain/stadium/CourtType.java
@@ -1,4 +1,4 @@
-package com.example.tuktuk.domain.court;
+package com.example.tuktuk.domain.stadium;
 
 import lombok.Getter;
